backend/pyproject.toml
[project]
name = "indexxxer"
version = "0.1.0"
description = "Media indexing and search platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",  # form data / file uploads
    "aiofiles>=23.2",           # async file streaming

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.29",
    "alembic>=1.13",

    # Config / validation
    "pydantic>=2.7",
    "pydantic-settings>=2.3",

    # Queue
    "celery[redis]>=5.4",
    "redis>=5.0",

    # File watching
    "watchdog>=4.0",

    # Media extraction
    "ffmpeg-python>=0.2",
    "Pillow>=10.4",
    "exifread>=3.0",

    # HTTP client (Ollama, Typesense in M2+)
    "httpx>=0.27",

    # Logging
    "structlog>=24.1",

    # pgvector — SQLAlchemy Vector type + asyncpg codec
    "pgvector>=0.3",

    # Auth — JWT + password hashing
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=3.2,<4.0",       # passlib 1.7.x is incompatible with bcrypt 4+
    "email-validator>=2.0",   # required by pydantic EmailStr (schemas/user.py)

    # Connectors — SMB + FTP
    "smbprotocol>=1.12",
    "aioftp>=0.22",
    "cryptography>=42",

    # GraphQL
    "strawberry-graphql[fastapi]>=0.235",

    # PDF rendering
    "pymupdf>=1.24",
]

[project.optional-dependencies]
# Install with: uv sync --extra ml
# Required on the GPU worker for CLIP, BLIP-2, Whisper, InsightFace
ml = [
    "open-clip-torch>=2.26",
    "faster-whisper>=1.1",
    "insightface>=0.7",
    "onnxruntime-gpu>=1.18",
    "transformers>=4.44",
    "accelerate>=0.34",
    "opencv-python-headless>=4.10",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "factory-boy>=3.3",
    "coverage[toml]>=7",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "alembic/*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
